Moving dagger modules to /modules
